+++ packages/backend/routes/collections.ts
@@ -421,8 +421,9 @@ export const collections: FastifyPluginAsync = async (
       preHandler: verifyCollectionOwner,
     },
     async (request, reply) => {
-      const { params } = request;
-      const { id, itemId } = params;
+      const {
+        params: { id, itemId },
+      } = request;
 
       try {
         const details = await pool.one(getItemDetails(id, itemId));
@@ -436,6 +437,8 @@ export const collections: FastifyPluginAsync = async (
           thumbnail_url,
           reading_time,
           collection_id,
+          collection_title,
+          collection_icon,
           ...rest
         } = details;
 
@@ -449,6 +452,11 @@ export const collections: FastifyPluginAsync = async (
           dateUpdated: timestampMsToSeconds(date_updated),
           thumbnailUrl: thumbnail_url,
           readingTime: reading_time,
+          collection: {
+            id: collection_id,
+            title: collection_title,
+            icon: collection_icon,
+          },
         };
       } catch (error) {
         if (error instanceof NotFoundError) {
@@ -462,6 +470,9 @@ export const collections: FastifyPluginAsync = async (
           logger.error(
             'There is more than one row matching the select criteria.'
           );
+          reply.status(500).send({ errorCode: 500, message: 'Server error.' });
+        } else {
+          logger.error(error);
           reply.status(500).send({ errorCode: 500, message: 'Server error.' });
         }
       }