+++ packages/frontend/components/article/ArticleHeader.stories.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Meta, Story } from '@storybook/react';
 import { ArticleHeader, ArticleHeaderProps } from './ArticleHeader';
 
@@ -15,6 +16,9 @@ const article = {
   title: 'Fun Article',
   datePublished: 1633781700,
   readingTime: 14.5,
+  collection: {
+    title: 'Fun One',
+  },
 } as any;
 
 export const Default = Template.bind({});
@@ -53,5 +57,20 @@ LongWord.args = {
   },
 };
 LongWord.parameters = {
+  ...Mobile.parameters,
+};
+
+export const LongCollection = Template.bind({});
+LongCollection.args = {
+  ...Default.args,
+  article: {
+    ...article,
+    collection: {
+      ...article.collection,
+      title: 'killbutmakeitlooklikeanaccident.sh',
+    },
+  },
+};
+LongCollection.parameters = {
   ...Mobile.parameters,
 };